Renamed OptionCommand to SwitchCommand and changed it to only work with commands
